requirements.txt
setuptools==57.5.0
django-celery
wheel==0.37.0
gunicorn==19.7.1
amqp==5.1.1
antiorm==1.2.1
asgiref==3.6.0
async-timeout==4.0.3
beautifulsoup4==4.11.2
billiard==3.6.4.0
bleach==6.0.0
celery
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.4
db==0.1.1
db-sqlite3==0.0.1
defusedxml==0.7.1
dj-rest-auth==5.0.1
Django==4.1.6
django-allauth==0.56.1
django-bootstrap4==22.3
django-filter==22.1
django-icons==22.1
django-summernote==0.8.20.0
django-taggit==3.1.0
django-tof==0.3.2
djangorestframework==3.14.0
idna==3.4
kombu==5.3.2
Markdown==3.4.4
oauthlib==3.2.2
Pillow==9.4.0
prompt-toolkit==3.0.39
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tzdata==2022.7
urllib3==2.0.5
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
whitenoise==6.6.0

